Allow setting trusted certs for OTLP Trace exporter
